feat: resetInstanceChild関数と関連する補助関数を新規作成
- resetInstanceChild関数を実装し、ancestorInstanceのオーバーライドをリセットする処理を追加
- getOverrideValues関数を新規作成し、オーバーライドの値を取得するロジックを実装
- restoreBoundVariables、restoreComponentProperties、restoreStyledTextSegment関数を新規作成し、各種プロパティの復元処理を実装
